joybus: move eeprom functions in different module.
No functional changes.

Also reorganize a bit the documentation groups in doxygen creating
a Peripherals section.
