+++ include/joybus.h
@@ -1,7 +1,7 @@
 /**
  * @file joybus.h
  * @brief Joybus Subsystem
- * @ingroup joybus
+ * @ingroup lowlevel
  */
 #ifndef __LIBDRAGON_JOYBUS_H
 #define __LIBDRAGON_JOYBUS_H
@@ -10,7 +10,7 @@
 #include <stdint.h>
 
 /**
- * @addtogroup joybus
+ * @addtogroup lowlevel
  * @{
  */
 
@@ -23,36 +23,12 @@
  */
 #define JOYBUS_BLOCK_DWORDS ( JOYBUS_BLOCK_SIZE / sizeof(uint64_t) )
 
-/**
- * @brief EEPROM Probe Values
- * @see #eeprom_present
- */
-typedef enum eeprom_type_t
-{
-    /** @brief No EEPROM present */
-    EEPROM_NONE = 0,
-    /** @brief 4 kilobit (64-block) EEPROM present */
-    EEPROM_4K   = 1,
-    /** @brief 16 kilobit (256-block) EEPROM present */
-    EEPROM_16K  = 2
-} eeprom_type_t; 
-
-/**
- * @brief Size of an EEPROM save block in bytes.
- */
-#define EEPROM_BLOCK_SIZE 8
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
 void joybus_exec( const void * inblock, void * outblock );
-eeprom_type_t eeprom_present( void );
-size_t eeprom_total_blocks( void );
-void eeprom_read( uint8_t block, uint8_t * dest );
-uint8_t eeprom_write( uint8_t block, const uint8_t * src );
-void eeprom_read_bytes( uint8_t * dest, size_t start, size_t len );
-void eeprom_write_bytes( const uint8_t * src, size_t start, size_t len );
 
 #ifdef __cplusplus
 }
