+++ include/libdragon.h
@@ -36,6 +36,7 @@
 #include "display.h"
 #include "dma.h"
 #include "dragonfs.h"
+#include "eeprom.h"
 #include "eepromfs.h"
 #include "graphics.h"
 #include "interrupt.h"
